instrumentation for tern_prog_begin and tern_prog_end
removed the use of atexit()

implemented instrumetnation code to insert tern_prog_begin and
tern_prog_end as static constructor and destructor.  Note both have to be
inserted as the last element in the LLVM static constructor or desctructor
list, so that tern_prog_begin is the first static ctor to be invoked, and
tern_prog_end is the last static dtor to be invoked at runtime.  (WHY?)