RaceDetector speed optimization
Each time we add a new access, we scan the read or write access list to
prune the ones that happen-before this new access.  This both saves
storage, as well as reduce the time needed to detect races.

This worked fine for the race detector in old tern.  However, our
happens-before now is strict happens-before, s.t. if two accesses are from
the same instruction trunk, the earlier instruction won't happen-before
the later one.  Thus, while we add accesses from an instruction trunk,
these accesses would pile up since later ones won't remove earlier ones.
Scanning this long list thus takes a lot of time.

To fix this problem, now we key @reads and @writes on timestamp.  Instead
of scanning through the entire list of accesses, we scan through
timestamps and if the timestamp TS happens-before the timestamp of the new
access, we remove all accesses with timestamp TS.