changed how __tern_prog_begin() and __tern_prog_end() are inserted
We used to insert __tern_prog_begin() as a static constructor.  However,
this doesn't work well with newer versions of gcc and binutils.  The
symptom is that __tern_prog_begin() is called before static constructors
in our runtime libraries are called, causing segv and other errors.

Since it is hard to anticipate how gcc/binutils will change the order of
static constructors, let's revert back to the approach to wrap
__tern_prog_begin() and __tern_prog_end() around the user's main() method.

In addition, __tern_prog_begin() registers __tern_prog_end() with
atexit(), so we only need to insert __tern_prog_begin() before main().

This change applies to both static and dynamic instrumentation.